Move search nodes to separate file
